refactor(extension): CHECKOUT-7635 Improve Command Handler Scalability
