+++ packages/checkout-extension/src/ExtensionService.ts
@@ -1,7 +1,8 @@
-import { CheckoutService, ExtensionCommandType, ExtensionRegion } from '@bigcommerce/checkout-sdk';
+import { CheckoutService, Extension, ExtensionRegion } from '@bigcommerce/checkout-sdk';
 import React from 'react';
 
-import { ExtensionAction, ExtensionActionType } from './ExtensionProvider';
+import { ExtensionAction } from './ExtensionProvider';
+import * as commandHandlers from './handler';
 
 export class ExtensionService {
     private handlers: { [extensionId: string]: Array<() => void> } = {};
@@ -10,10 +11,6 @@ export class ExtensionService {
         private checkoutService: CheckoutService,
         private dispatch: React.Dispatch<ExtensionAction>,
     ) {}
-
-    createAction(action: ExtensionAction): void {
-        this.dispatch(action);
-    }
 
     async loadExtensions(): Promise<void> {
         await this.checkoutService.loadExtensions();
@@ -28,55 +25,7 @@ export class ExtensionService {
 
         await this.checkoutService.renderExtension(container, region);
 
-        if (!this.handlers[extension.id]) {
-            this.handlers[extension.id] = [];
-        }
-
-        // TODO: CHECKOUT-7635
-        this.handlers[extension.id].push(
-            this.checkoutService.handleExtensionCommand(
-                extension.id,
-                ExtensionCommandType.ReloadCheckout,
-                () => {
-                    void this.checkoutService.loadCheckout(
-                        this.checkoutService.getState().data.getCheckout()?.id,
-                    );
-                },
-            ),
-        );
-
-        this.handlers[extension.id].push(
-            this.checkoutService.handleExtensionCommand(
-                extension.id,
-                ExtensionCommandType.SetIframeStyle,
-                (data) => {
-                    const { style } = data.payload;
-                    const extensionContainer = document.querySelector(
-                        `div[data-extension-id="${extension.id}"]`,
-                    );
-                    const iframe = extensionContainer?.querySelector('iframe');
-
-                    if (iframe) {
-                        Object.assign(iframe.style, style);
-                    }
-                },
-            ),
-        );
-
-        this.handlers[extension.id].push(
-            this.checkoutService.handleExtensionCommand(
-                extension.id,
-                ExtensionCommandType.ShowLoadingIndicator,
-                (data) => {
-                    const { show } = data.payload;
-
-                    this.createAction({
-                        type: ExtensionActionType.SHOW_LOADING_INDICATOR,
-                        payload: show,
-                    });
-                },
-            ),
-        );
+        this.registerHandlers(extension);
     }
 
     removeListeners(region: ExtensionRegion): void {
@@ -103,5 +52,23 @@ export class ExtensionService {
         const extension = this.checkoutService.getState().data.getExtensionByRegion(region);
 
         return Boolean(extension);
+    }
+
+    private registerHandlers(extension: Extension): void {
+        const handlerProps = {
+            checkoutService: this.checkoutService,
+            dispatch: this.dispatch,
+            extension,
+        };
+
+        if (!this.handlers[extension.id]) {
+            this.handlers[extension.id] = [];
+        }
+
+        Object.values(commandHandlers).forEach((handler) => {
+            if (typeof handler === 'function') {
+                this.handlers[extension.id].push(handler(handlerProps));
+            }
+        });
     }
 }
