+++ packages/checkout-extension/src/handler/ReloadCheckout.ts
@@ -1,0 +1,13 @@
+import { ExtensionCommandType } from '@bigcommerce/checkout-sdk';
+
+import { CommandHandler } from './CommandHandler';
+
+export const reloadCheckout: CommandHandler = ({ checkoutService, extension }) => {
+    return checkoutService.handleExtensionCommand(
+        extension.id,
+        ExtensionCommandType.ReloadCheckout,
+        () => {
+            void checkoutService.loadCheckout(checkoutService.getState().data.getCheckout()?.id);
+        },
+    );
+};
