+++ packages/checkout-extension/src/handler/ShowLoadingIndicator.ts
@@ -1,0 +1,20 @@
+import { ExtensionCommandType } from '@bigcommerce/checkout-sdk';
+
+import { ExtensionActionType } from '../ExtensionProvider';
+
+import { CommandHandler } from './CommandHandler';
+
+export const ShowLoadingIndicator: CommandHandler = ({ checkoutService, dispatch, extension }) => {
+    return checkoutService.handleExtensionCommand(
+        extension.id,
+        ExtensionCommandType.ShowLoadingIndicator,
+        (data) => {
+            const { show } = data.payload;
+
+            dispatch({
+                type: ExtensionActionType.SHOW_LOADING_INDICATOR,
+                payload: show,
+            });
+        },
+    );
+};
